CIF-747 - Shopping cart pop-up
* Fix the syntax so that it also works on Safari.
Safari and FF don't support some recent ES6 features like static properties on classes (methods implemented as arrow functions are considered properties as well)
* Wire the minicart body to the Commerce API, i.e. retrieve the products from Magento
I added a small piece of code that automatically creates the shopping cart when the page is loaded. This logic should be moved in the `addItem()` method.