CIF-747 - Shopping cart popup
The MiniCart is a full client-side component, implemented using plain Javascript. It relies on a CommerceApi implementation that provides the connection to the commerce backend.

The shopping cart is created automatically when the page loads and its details are saved in a cookie called `cif.cart`. Once the "Add to cart" action is implemented the cart will be created when the first product is added.