CIF-1436 - Can't remove or edit items from shopping cart (#314)
* CIF-1436 - Can't remove or edit items from shopping cart

The problem was caused by the minicart refactoring that we've done some time ago - when the first product is added to the cart the "handlers" were not initialized. I removed the "cart handler" methods and use the actions directly.

The same problem affects the coupons feature - both "Add coupon" and "Remove coupon" were broken because of the same issue. The same fix has been applied to those as well.

* Bring back the `remove coupon` and `add coupon` unit tests

* Add unit tests to increase coverage

Co-authored-by: Daniel Platon <platon@adobe.com>