CIF-1516 Use a short-lived resolver in ComponentsConfigurationAdapterFactory
Use the `try-with-resource` pattern to get a service resource resolver, so that the resolver is closed automatically.

The `ComponentsConfiguration` object has been updated so that internally it creates a new ValueMap. This is necessary because reusing the same ValueMap as the one returned by `asValueMap()` method is not possible when the resolver is closed.