CIF-2060: fixed Launches special handling of ComponentsConfiguration for newly created Resources
Launches do not fully support context aware configurations in all of the supported product versions. A workaround was introduced before to handle this special case in StoreConfigExporterImpl and MagentoGraphqlClient. In order to make those consumers agnostic to the implementation details of the workaround, the logic was moved to the ComponentsConfigurationAdapterFactory.